--- assurapp-api/src/test/java/be/helha/assurappapi/AssurappApiApplicationTests.java
@@ -1,13 +1,0 @@
-package be.helha.assurappapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AssurappApiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
